Adds Drango RestFramework functionality
